+++ src/tests/player/util.c
@@ -11,6 +11,7 @@ int setup_tests(void **state) {
 	z_info = mem_zalloc(sizeof(struct angband_constants));
 	z_info->pack_size = 23;
 	z_info->quest_max = 1;
+	z_info->quest_fixed = 1;
 	z_info->quiver_size = 10;
 	quests = &test_quest;
 	player_init(p);